+++ app/src/main/java/com/kalex/dogescollection/authentication/createaccount/presentation/CreateAccountViewModel.kt
@@ -2,7 +2,7 @@ package com.kalex.dogescollection.authentication.createaccount.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.kalex.dogescollection.authentication.AuthenticationUseCase
+import com.kalex.dogescollection.authentication.model.AuthenticationUseCase
 import com.kalex.dogescollection.authentication.createaccount.dto.User
 import com.kalex.dogescollection.common.networkstates.UseCaseFlowStatus
 import com.kalex.dogescollection.common.networkstates.ViewModelNewsUiState
