+++ app/src/main/java/com/kalex/dogescollection/authentication/model/AuthenticationUseCase.kt
@@ -1,4 +1,4 @@
-package com.kalex.dogescollection.authentication
+package com.kalex.dogescollection.authentication.model
 
 import com.kalex.dogescollection.authentication.createaccount.dto.User
 import com.kalex.dogescollection.authentication.createaccount.dto.userDTOtoUser
